refactor(config): move beckn constants into pkg/beckndefaults
- rename pkg/config → pkg/beckndefaults for a purpose-specific name;
  pkg/config was too generic and implied it was the adapter's general
  config package
- fold cmd/adapter/beckndefaults.go into cmd/adapter/main.go; the
  separate file added no reusability and split related logic without
  a clear boundary
- update remote refresh URLs to reflect new package path
